fix(processing): make output_cap.rs rustfmt-stable with real ratchet headroom
rustfmt --edition 2021 --check disagreed with the committed output_cap.rs:
it wanted to reflow the four new try_push(...) call sites (each 87-99
columns) into 3-line closures plus a fifth pre-existing struct literal,
which would land the file at ~408 lines against the 400-line
module_size_ratchet ceiling the moment anyone ran cargo fmt.

Apply the formatting rustfmt actually wants, and move
SymbolicTruncationReason/SymbolicTruncation -- self-contained wire types
with no dependency on the accumulator's push/charge logic -- into a new
sibling module, output_cap_types.rs, so there is real slack rather than
landing exactly at the limit again. No allowlist edit.

output_cap.rs: 326 lines, output_cap_validate.rs: 42 lines,
output_cap_types.rs: 97 lines. Both output_cap.rs and
output_cap_validate.rs are fmt-clean under rustfmt --edition 2021 --check.
