+++ rust/processing/src/symbolic/output_cap.rs
@@ -11,12 +11,12 @@
 //! to find it, and a reader adding a primitive should not have to step around
 //! the bound.
 
+use super::output_cap_types::{SymbolicTruncation, SymbolicTruncationReason};
 use super::output_cap_validate::all_finite;
 use super::primitives::{
     SymbolicCircle, SymbolicData, SymbolicFillArea, SymbolicGridAxis, SymbolicPolyline,
     SymbolicText,
 };
-use serde::{Deserialize, Serialize};
 
 /// Upper bound on the total number of symbolic primitives one extraction may
 /// emit, across every product in the file.
@@ -82,91 +82,6 @@ pub const MAX_SYMBOLIC_BYTES: usize = 256 * 1024 * 1024;
 const PRIMITIVE_OVERHEAD_BYTES: usize = 64;
 /// Charged per `f32` coordinate or per byte of text.
 const BYTES_PER_COORD: usize = 8;
-
-/// Which bound stopped an extraction early.
-///
-/// `SymbolicData` had no diagnostics channel at all (#2938), so a drawing that
-/// lost 60% of its curves was indistinguishable, in the response, from one that
-/// legitimately had nothing more to emit.
-///
-/// The reason matters as much as the fact. #2938's own lead case is a
-/// well-formed nested block import losing content to the PER-ITEM revisit
-/// budget while the whole-file totals sit far below the extraction bounds --
-/// so a diagnostic that only reported the extraction bounds would have reported
-/// nothing on the exact scenario the issue is about.
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(rename_all = "kebab-case")]
-pub enum SymbolicTruncationReason {
-    /// [`MAX_SYMBOLIC_ELEMENTS`] reached.
-    ElementCount,
-    /// [`MAX_SYMBOLIC_BYTES`] reached.
-    OutputBytes,
-    /// One representation item nested deeper than the walk follows.
-    ItemDepth,
-    /// The EXTRACTION exhausted its revisit budget: a large acyclic fan-out,
-    /// or a legitimate deeply-nested block import.
-    ///
-    /// Shared across the whole file since #3114, not per item -- a per-item
-    /// budget reset on every top-level item, so nothing bounded a fan-out
-    /// spread across many items. The name is kept for wire compatibility.
-    /// Note the budget is extraction-wide while `ItemWalk::seen` stays per
-    /// item, so re-placing one library block is not charged as a revisit.
-    ItemRevisits,
-    /// The walk's path guard (`ItemWalk::enter_node`) refused to re-enter a
-    /// node already on the current path -- a genuine cycle in the
-    /// representation graph, not merely a large fan-out. Distinct from
-    /// [`Self::ItemRevisits`], whose budget can also be exhausted by an
-    /// acyclic file (#2938's lead case); this reason is a cycle and nothing
-    /// else (#3108).
-    ItemCycle,
-}
-
-impl SymbolicTruncationReason {
-    /// The wire spelling, identical to what `Serialize` emits.
-    ///
-    /// The WASM boundary cannot hand a serde enum to JavaScript, so it needs a
-    /// plain string; having it here rather than a `match` in wasm-bindings keeps
-    /// one vocabulary for both surfaces. `the_wire_spellings_match_serde` pins
-    /// them together, because two hand-kept lists is how they drift.
-    pub fn as_wire_str(self) -> &'static str {
-        match self {
-            Self::ElementCount => "element-count",
-            Self::OutputBytes => "output-bytes",
-            Self::ItemDepth => "item-depth",
-            Self::ItemRevisits => "item-revisits",
-            Self::ItemCycle => "item-cycle",
-        }
-    }
-}
-
-/// What stopped an extraction early, when something did.
-///
-/// Present only on a truncated result.
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-pub struct SymbolicTruncation {
-    /// The MOST SEVERE bound that fired, not the first: an extraction bound
-    /// outranks a per-item one whatever the scan order. See
-    /// `SymbolicAccumulator::record`.
-    pub reason: SymbolicTruncationReason,
-    /// Primitives emitted in total. NOT necessarily equal to any limit: a
-    /// traversal bound stops content from being produced while the file-level
-    /// totals stay far below the extraction bounds.
-    pub emitted: usize,
-    /// The bound's value, when the reason has a single numeric one. `None` for
-    /// the traversal reasons, whose bounds count a DIFFERENT UNIT from
-    /// `emitted` -- revisits and nesting depth, not primitives -- so there is
-    /// no meaningful "{emitted} of {limit}" to render.
-    ///
-    /// Note this is no longer because those bounds are per item: since #3114
-    /// the revisit budget is extraction-wide. It is the units that do not
-    /// line up, and that is what keeps `limit` absent.
-    ///
-    /// Skipped rather than serialized as `null`: the TypeScript mirror declares
-    /// `limit?: number`, which means the key is ABSENT. Emitting `null` satisfies
-    /// Rust and breaks the consumer's `'limit' in truncated` check.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub limit: Option<usize>,
-}
 
 /// The extraction's accumulator: a `SymbolicData` under construction, plus the
 /// bound it is being built under.
@@ -332,15 +247,18 @@ impl SymbolicAccumulator {
         // file-bounded so it has no fan-out amplifier, but an uncharged
         // heap field is the same class of hole as `alignment` was.
         let payload = axis.tag.len();
-        self.try_push(payload, all_finite(&axis.endpoints), |data| data.grid_axes.push(axis));
+        self.try_push(payload, all_finite(&axis.endpoints), |data| {
+            data.grid_axes.push(axis)
+        });
     }
 
     /// Append a polyline unless the extraction has hit its cap.
     pub(super) fn push_polyline(&mut self, polyline: SymbolicPolyline) {
-        let payload = polyline.points.len()
-            + polyline.ifc_type.len()
-            + polyline.representation.len();
-        self.try_push(payload, all_finite(&polyline.points), |data| data.polylines.push(polyline));
+        let payload =
+            polyline.points.len() + polyline.ifc_type.len() + polyline.representation.len();
+        self.try_push(payload, all_finite(&polyline.points), |data| {
+            data.polylines.push(polyline)
+        });
     }
 
     /// Append a circle unless the extraction has hit its cap.
@@ -367,7 +285,9 @@ impl SymbolicAccumulator {
             + text.alignment.len()
             + text.ifc_type.len()
             + text.representation.len();
-        self.try_push(payload, all_finite(&[text.x, text.y]), |data| data.texts.push(text));
+        self.try_push(payload, all_finite(&[text.x, text.y]), |data| {
+            data.texts.push(text)
+        });
     }
 
     /// Append a filled region unless the extraction has hit its cap.
@@ -376,7 +296,9 @@ impl SymbolicAccumulator {
             + fill.holes_offsets.len()
             + fill.ifc_type.len()
             + fill.representation.len();
-        self.try_push(payload, all_finite(&fill.points), |data| data.fills.push(fill));
+        self.try_push(payload, all_finite(&fill.points), |data| {
+            data.fills.push(fill)
+        });
     }
 
     /// Finish, stamping the diagnostics field iff an append was ever refused.
@@ -393,7 +315,11 @@ impl SymbolicAccumulator {
                 | SymbolicTruncationReason::ItemRevisits
                 | SymbolicTruncationReason::ItemCycle => None,
             };
-            self.data.truncated = Some(SymbolicTruncation { reason, emitted, limit });
+            self.data.truncated = Some(SymbolicTruncation {
+                reason,
+                emitted,
+                limit,
+            });
         }
         self.data
     }
